Remove Console's optional dependency on Application
This is because Applications requires Console, so Console needs to be
instantiated first.
